Split out overlay
